+++ samples/shared/src/commonMain/kotlin/com/github/panpf/zoomimage/sample/ui/base/ToolbarScaffold.kt
@@ -4,10 +4,13 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.BoxScope
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.WindowInsets
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.navigationBars
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material3.ExperimentalMaterial3Api
@@ -24,6 +27,7 @@ import com.github.panpf.zoomimage.sample.ui.LocalNavBackStack
 @Composable
 fun ToolbarScaffold(
     title: String,
+    addNavigationBarWindowInsets: Boolean = false,
     content: @Composable BoxScope.() -> Unit,
 ) {
     val navBackStack = LocalNavBackStack.current
@@ -46,7 +50,14 @@ fun ToolbarScaffold(
                 )
             }
         )
-        Box(Modifier.fillMaxWidth().weight(1f)) {
+        Box(
+            Modifier.fillMaxWidth()
+                .weight(1f)
+                .let {
+                    if (addNavigationBarWindowInsets)
+                        it.windowInsetsPadding(WindowInsets.navigationBars) else it
+                }
+        ) {
             content()
         }
     }