+++ samples/shared/src/desktopMain/kotlin/com/github/panpf/zoomimage/sample/ui/preview/ZoomImageToolPreview.kt
@@ -33,6 +33,6 @@ fun ZoomImageToolPreview() {
         infoDialogState = rememberMyDialogState(),
         capturableState = rememberCapturableState(),
         photoPaletteState = remember { mutableStateOf(PhotoPalette(colorScheme)) },
-        scrollBar = ScrollBarSpec.DefaultAndWindowInsets
+        scrollBar = ScrollBarSpec.MediumAndWindowInsets
     )
 }