+++ CHANGELOG.md
@@ -9,6 +9,8 @@ Translations: [简体中文](CHANGELOG.zh.md)
 * new: The scroll bar now supports WindowInsets, which can be turned on through the
   enabledWindowInsets parameter of ScrollBarSpec. Please refer to the documentation for
   details. [scrollbar.md](docs/scrollbar.md)
+* new: ScrollBarSpec adds DefaultAndWindowInsets, Medium, MediumAndWindowInsets, Large,
+  LargeAndWindowInsets configuration templates
 
 ## Android